Improve HTTP::Tiny 599 Internal Exception diagnostic
The correct thing to do for packages that require https is to use a https: start url,
or failing that to use the ssl property on the download or fetch plugin that you use.
The error that you get for failing to do that is very confusing, so I am expanding
on the diagnostic, and adding a FAQ item on the issue.