Hardcode CMAKE_INSTALL_LIBDIR as lib
This does for CMake the same 31115d0 ("use config.site to ensure lib
instead of lib64 on Fedora") did for Autoconf:
It tries to ensure that the pkg-config and library files are
installed into lib/, not lib64 or some other OS-specific subdir.

This change is necessary to support projects configured using CMake
and installing into GNUInstallDirs provided paths on systems like
Fedora 28 64-bit.