allow fail on 5.8.1 where Net::SSLeay is borked.
Net::SSLeay is essentially broken on 5.8.7 and earlier.
IO::Socket::SSL is probably not good either.  Long term we want
to prefer curl if ALREADY installed and if it supports https
over these two modules when they aren't ALREADY installed.
For now we will allow it to fail until we can fix this.