try a couple of times to get the index via curl
curl -O -J is what we want for tarballs, etc because we get the right filename.
Unfortunately, it mucks up URLs with trailing '/', because it assumes the filename
is '', which can't be written to.  Checking the URL for a trailing '/' isn't good
enough, because there may be redirects involved.

So to work around this, try once with '-O -J' and once with '-o index.html' if
we see the daignostic that indicates an empty filename.