Avoid C99-related errors in expected-compiler-error test
This commit supplies the standard XS includes, to avoid triggering
implicit ints and implicit functions in the standard code fragments
generated by ExtUtils::ParseXS and related tools.

For this test case, this is not a real problem because the
compiliation is expected to fail anyway, but the presence of some
C99-related errors could mask other issues in the package during
future testing.